requirements.txt
# MongoDB GraphRAG Infrastructure Demo
# Requirements

# Core
pymongo[srv]==4.6.1
motor==3.3.2

# LangChain ecosystem
langchain==0.3.13
langchain-community==0.3.13
langchain-mongodb==0.3.0
langchain-anthropic==0.3.0

# Embeddings (VoyageAI)
voyageai==0.2.3

# LLM (Anthropic)
anthropic==0.40.0

# Vector operations
numpy==1.26.4

# Environment & Config
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Demo UI
streamlit==1.33.0
plotly==5.22.0

# Data handling
pandas==2.2.2

# Development
pytest==8.2.0
pytest-asyncio==0.23.6
rich==13.7.1

